[polaris.shopify.com] Rename AlphaStack and update examples
